0866002 lab5

+++ test/preemption.c
@@ -3,7 +3,7 @@
 #include <timer.h>
 
 static void
-delay (int sec)
+delay (size_t sec)
 {
   size_t t, cnt, freq;
   sys_get_time (&cnt, &freq);
